% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get.pseudo.obs}
\alias{get.pseudo.obs}
\title{get.pseudo.obs}
\usage{
get.pseudo.obs(ticker, u)
}
\arguments{
\item{ticker}{ticker symbol for the asset}

\item{u}{chosen threshold for the upper tail}
}
\value{
numeric vector of uniform pseudo-observations
}
\description{
get.pseudo.obs
}
