% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{t.dist.plot}
\alias{t.dist.plot}
\title{fit t distribution plot}
\usage{
\method{t}{dist.plot}(ticker)
}
\arguments{
\item{data}{data}
}
\value{
ggplot2
}
\description{
function for fitting t distribution and returning plot
}
